siteA/siteA/requirements.txt
Django>=6,<6.1
wagtail==7.3rc1
pillow-avif-plugin
